rake compile task

+++ lib/ffi_test.c
@@ -1,18 +1,18 @@
 #include <stdlib.h>
 #include <math.h>
 
-int ffi_pow(int a, int n) {
+unsigned long long int ffi_pow(unsigned long long int a, unsigned long long int n) {
   return pow(a, n);
 }
 
-int ffi_factorial(int max) {
-  int i=max, result=1;
+unsigned long long int ffi_factorial(unsigned long long int max) {
+  unsigned long long int i=max, result=1;
   while (i >= 2) { result *= i--; }
   return result;
 }
 
-int ffi_fibonacci(int n) {
-  int a = 1, b = 1, c, i;
+unsigned long long int ffi_fibonacci(unsigned long long int n) {
+  unsigned long long int a = 1, b = 1, c, i;
   if (n == 0) {
     return 0;
   }
